+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-06-03 16:31:17
-// Commit: fix(security): prevent double tap duplicate action in security
+// Updated at: 2026-06-04 12:19:56
+// Commit: feat(security): add bi-directional event bus handler for security
 
 class securityComponentState {
-  final String commitRef = '1713';
-  final String lastMessage = 'fix(security): prevent double tap duplicate action in security';
-  final DateTime timestamp = DateTime.parse('2026-06-03T16:31:17+05:00');
+  final String commitRef = '1744';
+  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
+  final DateTime timestamp = DateTime.parse('2026-06-04T12:19:56+05:00');
 }